requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
email-validator==2.1.0.post1
werkzeug==2.3.7
twilio==8.10.2
python-dotenv==1.0.0
